Tests galore!
There are now tests of the code base with test coverage. It covers most of the
code except for the websocket interactions with the TV.

As part of this some minor bugs were found which has been corrected. Alga also
gives a non-zero exit code in case of errors.

Dependencies have also been updated.