Specify minimum versions for runtime dependencies (#5)
This means Alga doesn't just allow any version of the dependencies to be
installed, as they might not have the necessary functionality in the older
versions.

It would be ideal to run tests against these lower ranges in CI to make sure
they keep working. That would currently require using uv or PDM as they support
installing the lowest version of packages that are allowed inside the range. It
can be looked into later.