Write details to a config file when setting up a TV connection
This means environment variables no longer are used for configuring Alga, and
means `alga power on` no longer takes an argument since the MAC address now is
stored in the config file.

This means Alga currently only supports one TV at a time, but I assume that's
fine for the vast majority of users. If requests come in for support for
multiple TVs, then it can be added, but for now it reduces the complexity.