Set TV channel using the channel ID instead of channel number
Some models of TVs, e.g. the 50UN73006LA requires the `ssap://tv/openChannel`
call to be provided with a channel ID instead of a channel number.

Since the channel ID is something obscure like "1_21_101_101_16515_17539_9018"
it doesn't make sense for a user to try to input the ID manually. Instead this
change makes the `channel set` command look up the list of channels to convert
a channel number to a channel ID, if the input looks like a channel number.