Add support for managing multiple TVs
This adds the following commands for managing the TVs in the configuration:

* `alga tv add` (replacing `alga setup`)
* `alga tv list`
* `alga tv remove`
* `alga tv rename`
* `alga tv set-default`

A new global option has also been added, which allows specifying which TV to
send commands to on a command-by-command basis. It has to be used like `alga
--tv <identifer> ...`.

The WebSocket client code has been refactored as part of this, to make it
easier to split up the use cases of adding a new TV (where a handshake
shouldn't be performed) and sending regular commands to manage a TV. Tests has
also been added for this part of the code.

A new version of the configuration file format has been added and Alga will
automatically migrate to it if an old configuration file format is detected.
When this is done, the existing TV in the configuration will be assigned an
identifier for "default".