Switch to use uv instead of Poetry (#175)
* Replace Poetry configuration with uv

* Update commands in README to point at running uv

* Adapt checks workflow to use uv

This also removes the `DeterminateSystems/magic-nix-cache-action` action which
no longer works and adds the "pytest-lowest" job which uses uv's feature to
resolve dependencies to their lowest possible version, so it can be verified it
works with those as well.

* Adapt release workflow to use uv

* Merge pytest-lowest job into existing pytest job

This should also fix the reference to Python 3.14 and disable the cache of
dependencies when running tests, because the cache otherwise will collide.

* Lock dev dependencies to their latest versions

We don't want them to be using their oldest versions when running tests with
`--resolution=lowest`.

* Cobble together a flake.nix file that seems to work with uv