Run test command with the `--help` argument
Version 8.2.0 of Click set the exit code to 2 when a command is run without any
arguments, causing the help output to be outputted, but not if `--help` is
passed.

The CI checks started failing when Click was updated due to this, so adding the
extra flag means the exit code returns to 0 again.